+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Snowpark Python API Updates
 
+#### New Features
+
+- Added support for the following `xpath` functions in `functions.py`:
+  - `xpath`
+  - `xpath_string`
+  - `xpath_boolean`
+  - `xpath_int`
+  - `xpath_float`
+  - `xpath_double`
+  - `xpath_long`
+  - `xpath_short`
+
 #### Bug Fixes
 
 - Fixed a bug that `to_pandas()` has different format of column name when query result format is set to 'JSON' and 'ARROW'.